- Updated the xform_model.py script to run the tflite_micro_compiler and autogenerate the .cpp file. - Moved wrapper.cpp and wrapper.h out of src/inference/model since model should only contain autogenerated files
