+++ modules/lib_vnr/CMakeLists.txt
@@ -2,7 +2,7 @@
 ## VNR Inference
 add_library(fwk_voice_module_lib_vnr_inference STATIC)
 
-file(GLOB_RECURSE VNR_INFERENCE_SOURCES src/inference/model/*.cpp src/inference/*.cc)
+file(GLOB_RECURSE VNR_INFERENCE_SOURCES src/inference/model/*.cpp src/inference/*.cc src/inference/*.cpp)
 
 target_sources(fwk_voice_module_lib_vnr_inference PRIVATE ${VNR_INFERENCE_SOURCES})
