+++ modules/lib_vnr/python/utils/xformer/requirements.txt
@@ -1,5 +1,6 @@
 xmos-ai-tools==0.1.8
 jinja2
-tensorflow==2.8.0
+tensorflow==2.8.0; platform_machine != 'arm64'
+tensorflow-macos==2.9.0; sys_platform == 'darwin' and platform_machine == 'arm64'
 protobuf==3.20.1  # https://github.com/protocolbuffers/protobuf/issues/10051
