+++ modules/lib_vnr/src/inference/wrapper.h
